+++ src/main/java/org/betonquest/betonquest/events/ClearEvent.java
@@ -40,7 +40,7 @@ public class ClearEvent extends QuestEvent {
         for (int i = 0; i < types.length; i++) {
             try {
                 types[i] = EntityType.valueOf(entities[i].toUpperCase(Locale.ROOT));
-            } catch (IllegalArgumentException e) {
+            } catch (final IllegalArgumentException e) {
                 throw new InstructionParseException("Entity type '" + entities[i] + "' does not exist", e);
             }
         }
@@ -62,9 +62,6 @@ public class ClearEvent extends QuestEvent {
         final Collection<Entity> entities = location.getWorld().getEntities();
         loop:
         for (final Entity entity : entities) {
-            if (!(entity instanceof LivingEntity)) {
-                continue;
-            }
             if (name != null && (entity.getCustomName() == null || !entity.getCustomName().equals(name))) {
                 continue;
             }